用户行为相关 0.08

+++ rjra-rdb/rjra-rdb-proxy/src/main/java/com/hdg/rjra/rdb/proxy/daoproxy/IUserApplyWorkProxy.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
  */
 public interface IUserApplyWorkProxy extends Serializable {
 
-    public Integer saveUserApplyWork(UserApplyWork userApplyWork);
+    public Long saveUserApplyWork(UserApplyWork userApplyWork);
 
     public UserApplyWork findUserApplyWorkByUserIdAndWorkId(Long userId, Long workId);
 }
