+++ rjra-rdb/rjra-rdb-proxy/src/main/java/com/hdg/rjra/rdb/proxy/daoproxy/IUserInviteUserProxy.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
  */
 public interface IUserInviteUserProxy extends Serializable {
 
-    public Integer saveUserInviteUser(UserInviteUser userInviteUser);
+    public Long saveUserInviteUser(UserInviteUser userInviteUser);
 
     public UserInviteUser findUserInviteUserByUserIdAndInviteUserId(Long userId, Long inviteUserId);
 }
